# api-implementation/server/common/api.yml
openapi: 3.0.0
info:
  title: Solace Platform API
  description: Solace Platform API
  version: 0.0.1
servers:
  - url: /v1
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  parameters:
    organization:
      name: organization
      description: the name of the organization (tenant) a resource belongs to
      in: path
      required: true
      schema:
        type: string
...
    api_name:
      name: api_name
      description: the name of the api
      in: path
      required: true
      schema:
        type: string
    api_domain_name:
      name: api_domain_name
      description: the name of the api domain
      in: path
      required: true
      schema:
        type: string
...
    page_size:
     name: pageSize
     in: query
     description: "Currently not honoured. The number of itesm to get per page. Min: 1 Max: 100"
     required: false
     schema: 
      type: integer
      format: int32
      default: 20
    page_number:
     name: pageNumber
     in: query
     description: "Currently not honoured. The page number to get. Min: 1"
     required: false
     schema: 
       type: integer
       format: int32
       default: 1
  schemas:
    ErrorResponse: 
     type: object
     properties: 
      message: 
       type: string
       example: "An error occurred"
      errorId: 
       type: string
       example: "123e4567-e89b-12d3-a456-426655440000"
      meta: 
       type: object
       additionalProperties: 
        type: object
        example: "'object': { 'field': 'description' }"
       example: "'object': { 'field': 'description' }"
    APIDomain: 
     description: "Event Portal API Domain information. Currently only provided as utiity to retrieve metadata from the Event Portal."
     required: 
      - enforceUniqueTopicNames
     type: object
     properties: 
      createdTime: 
       type: integer
       format: int64
       readOnly: true
       example: 1583706406
      updatedTime: 
       type: integer
       format: int64
       readOnly: true
       example: 1583706406
      createdBy: 
       type: string
       readOnly: true
       example: joe.bloggs
      changedBy: 
       type: string
       readOnly: true
       example: jack.black
      id: 
       type: string
       readOnly: true
       example: 12345678
      name: 
       type: string
       example: "Acme Rideshare"
      topicDomain: 
       type: string
       example: "com/solace"
      description: 
       type: string
       example: "API Domain created by the Solace PubSub+ Cloud API documentation"
      enforceUniqueTopicNames: 
       type: boolean
       example: true
      apis:           
       example:
        - Billing App
        - Driver App
       type: array
       readOnly: true
       items: 
        type: string
        readOnly: true
      type: 
       type: string
    APIListItem:
      description: "API List item (short info). Currently only provided as utiity to retrieve metadata from the Event Portal."
      type: object
      properties:
        name: 
          type: string
          example: Billing App
        apiDomain: 
          type: string
          example: Acme Rideshare
    API: 
     description: "Event Portal API information. Currently only provided as utiity to retrieve metadata from the Event Portal."
     required: 
      - apiDomainName
      - name
     type: object
     properties: 
      createdTime: 
       type: integer
       format: int64
       readOnly: true
       example: 1583706406
      updatedTime: 
       type: integer
       format: int64
       readOnly: true
       example: 1583706406
      createdBy: 
       type: string
       readOnly: true
       example: joe.bloggs
      changedBy: 
       type: string
       readOnly: true
       example: jack.black
      id: 
       type: string
       readOnly: true
       example: 12345678
      name: 
       maxLength: 60
       minLength: 1
       type: string
       example: "Billing App"
      version: 
       type: string
       readOnly: true
       example: "0.0.1"
      description: 
       type: string
       example: "API created by Solace PubSub+ Cloud documentation"
      apiDomainName: 
       type: string
       example: Acme Rideshare
      revisionComment: 
       maxLength: 255
       minLength: 0
       type: string
       writeOnly: true
       example: "Changed api name"
      revisionNumber: 
       type: integer
       format: int32
       readOnly: true
       example: 
        - 1
        - 2
      apiClass: 
       type: string
       example: solace
       enum: 
        - unspecified
        - kafka_connector
        - solace_connector
        - kafka_application
      type: 
       type: string
       readOnly: true
     discriminator: 
      propertyName: apiClass
...
security:
  - BasicAuth: []
...
paths:
...
  /{organization}/apis:
    parameters:
    - $ref: '#/components/parameters/organization'
    get:
     tags:
      - apis
     parameters:
     - $ref: '#/components/parameters/page_size'
     - $ref: '#/components/parameters/page_number'
     responses:
       '200':
         description: list of APIs
         content: 
           application/json:
             schema:
               type: array
               items:
                 type: string
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
  /{organization}/apis/{api_name}:
    parameters:
    - $ref: '#/components/parameters/organization'
    - $ref: '#/components/parameters/api_name'
    put:
      tags: 
      - apis
      requestBody: 
       description: "Async API specification"
       content: 
         "text/plain": 
           schema: 
            type: string    
      responses:
       '201':
         description: The async API specification
         content: 
           application/json:
             schema:
               type: string
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '422': 
         description: "AlreadyExists."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
    get:
     tags:
     - apis
     responses:
       '200':
         description:  async api sepc
         content: 
           application/json:
             schema:
               type: string
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
    patch:
     tags:
     - apis
     requestBody: 
       description: "async api specification"
       content: 
         "text/plain": 
           schema: 
            type: string    
     responses:
       '200':
         description: updated async api specification
         content: 
           text/plain:
             schema:
               type: string
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '422': 
         description: "AlreadyExists."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
    delete:
     tags:
     - apis
     responses:
       '204':
         description: No content
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '422': 
         description: "AlreadyExists."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
...
  /{organization}/event-portal/apis:
    parameters:
    - $ref: '#/components/parameters/organization'
    get:
     tags:
      - eventPortal
     description: Returns a list of apps, maps to underlying event portal api /api/v1/eventPortal/applications
     parameters:
     - $ref: '#/components/parameters/page_size'
     - $ref: '#/components/parameters/page_number'
     - name: apiDomainName
       schema:
         type: string
       example: Acme Rideshare
       description: Match only apis in the given api domain.
       in: query
       required: false
     responses:
       '200':
         description:  a list of the APIs
         content: 
           application/json:
             schema:
               type: array
               items: 
                 "$ref": "#/components/schemas/APIListItem"
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
  /{organization}/event-portal/apis/{api_name}:
    parameters:
    - $ref: '#/components/parameters/organization'
    - $ref: '#/components/parameters/api_name'
    get:
     tags:
     - eventPortal
     responses:
       '200':
         description: the API  corresponding to the name requested
         content: 
           application/json:
             schema:
               type: array
               items: 
                 "$ref": "#/components/schemas/API"
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
...
  /{organization}/event-portal/apiDomains:
    parameters:
    - $ref: '#/components/parameters/organization'
    get:
     tags:
      - eventPortal
     parameters:
     - $ref: '#/components/parameters/page_size'
     - $ref: '#/components/parameters/page_number'
     responses:
       '200':
         description: the API domain corresponding to the name requested
         content: 
           application/json:
             schema:
               type: array
               items:
                 $ref: '#/components/schemas/APIDomain'
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
  /{organization}/event-portal/apiDomains/{api_domain_name}:
    parameters:
    - $ref: '#/components/parameters/organization'
    - $ref: '#/components/parameters/api_domain_name'
    get:
     tags:
     - eventPortal
     responses:
       '200':
         description: the API domain corresponding to the name requested
         content: 
           application/json:
             schema:
               $ref: '#/components/schemas/APIDomain'
       '400': 
         description: "Bad Request."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
          "*/*": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '401': 
         description: "Unauthorized."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '403': 
         description: "Forbidden."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '404': 
         description: "Not Found."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '500': 
         description: "Internal Server Error."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '503': 
         description: "Service Unavailable."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"
       '504': 
         description: "Gateway Timeout."
         content: 
          "application/json": 
           schema: 
            "$ref": "#/components/schemas/ErrorResponse"